Rename GameInAssetContext.ts -> getGameInAssetContext.ts
